--- Pestis/Assets/Split_Recommended.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9e2e58dacbeb3e74cbc85834b34f3d64
